refactor(health_connector_hk_ios): Separate health record mapper files for each health record type
