+++ packages/health_connector_hk_ios/ios/health_connector_hk_ios/Sources/health_connector_hk_ios/mappers/health_record_mappers/BodyTemperatureRecordMappers.swift
@@ -1,0 +1,65 @@
+import Foundation
+import HealthKit
+
+// ==================== BODY TEMPERATURE RECORD MAPPERS ====================
+
+extension BodyTemperatureRecordDto {
+    /**
+     * Converts this DTO to a HealthKit `HKQuantitySample`.
+     *
+     * - Throws: An error if the quantity type cannot be created.
+     */
+
+    func toHealthKit() throws -> HKQuantitySample {
+        guard let type = HKQuantityType.quantityType(forIdentifier: .bodyTemperature) else {
+            throw NSError(
+                domain: "HealthConnectorError",
+                code: -1,
+                userInfo: [NSLocalizedDescriptionKey: "Failed to create body temperature quantity type"]
+            )
+        }
+
+        let quantity = HKQuantity(unit: .degreeCelsius(), doubleValue: temperature.value)
+        let date = Date(timeIntervalSince1970: TimeInterval(time) / 1000.0)
+
+        return HKQuantitySample(
+            type: type,
+            quantity: quantity,
+            start: date,
+            end: date, // Instant records have same start and end
+            device: metadata.toHealthKitDevice(),
+            metadata: metadata.toHealthKitMetadata()
+        )
+    }
+}
+
+extension HKQuantitySample {
+    /**
+     * Converts this HealthKit sample to a `BodyTemperatureRecordDto`.
+     *
+     * Returns `nil` if this sample is not a body temperature sample.
+     */
+
+    func toBodyTemperatureRecordDto() -> BodyTemperatureRecordDto? {
+        guard quantityType.identifier == HKQuantityTypeIdentifier.bodyTemperature.rawValue else {
+            return nil
+        }
+
+        let unit = HKUnit.degreeCelsius()
+        let value = quantity.doubleValue(for: unit)
+
+        let metadataDict = metadata ?? [:]
+        let zoneOffset = metadataDict.extractTimeZoneOffset(for: startDate)
+        
+        return BodyTemperatureRecordDto(
+            id: uuid.uuidString,
+            metadata: metadataDict.toMetadataDto(
+                source: sourceRevision.source,
+                device: device
+            ),
+            time: Int64(startDate.timeIntervalSince1970 * 1000),
+            temperature: TemperatureDto(unit: TemperatureUnitDto.celsius, value: value),
+            zoneOffsetSeconds: zoneOffset
+        )
+    }
+}
