+++ packages/health_connector_hk_ios/ios/health_connector_hk_ios/Sources/health_connector_hk_ios/mappers/health_record_mappers/DistanceRecordMappers.swift
@@ -1,0 +1,64 @@
+import Foundation
+import HealthKit
+
+// ==================== DISTANCE RECORD MAPPERS ====================
+
+extension DistanceRecordDto {
+    /**
+     * Converts this DTO to a HealthKit `HKQuantitySample`.
+     *
+     * - Throws: An error if the quantity type cannot be created.
+     */
+
+    func toHealthKit() throws -> HKQuantitySample {
+        guard let type = HKQuantityType.quantityType(forIdentifier: .distanceWalkingRunning) else {
+            throw NSError(
+                domain: "HealthConnectorError",
+                code: -1,
+                userInfo: [NSLocalizedDescriptionKey: "Failed to create distance quantity type"]
+            )
+        }
+
+        let quantity = distance.toHealthKit()
+        let startDate = Date(timeIntervalSince1970: TimeInterval(startTime) / 1000.0)
+        let endDate = Date(timeIntervalSince1970: TimeInterval(endTime) / 1000.0)
+
+        return HKQuantitySample(
+            type: type,
+            quantity: quantity,
+            start: startDate,
+            end: endDate,
+            device: metadata.toHealthKitDevice(),
+            metadata: metadata.toHealthKitMetadata(timeZone: TimeZone.current)
+        )
+    }
+}
+
+extension HKQuantitySample {
+    /**
+     * Converts this HealthKit sample to a `DistanceRecordDto`.
+     *
+     * Returns `nil` if this sample is not a distance walking/running sample.
+     */
+
+    func toDistanceRecordDto() -> DistanceRecordDto? {
+        guard quantityType.identifier == HKQuantityTypeIdentifier.distanceWalkingRunning.rawValue else {
+            return nil
+        }
+
+        let metadataDict = metadata ?? [:]
+        let zoneOffset = metadataDict.extractTimeZoneOffset(for: startDate)
+        
+        return DistanceRecordDto(
+            distance: quantity.toLengthDto(),
+            endTime: Int64(endDate.timeIntervalSince1970 * 1000),
+            id: uuid.uuidString,
+            metadata: metadataDict.toMetadataDto(
+                source: sourceRevision.source,
+                device: device
+            ),
+            startTime: Int64(startDate.timeIntervalSince1970 * 1000),
+            zoneOffsetSeconds: zoneOffset
+        )
+    }
+}
