+++ packages/health_connector_hk_ios/ios/health_connector_hk_ios/Sources/health_connector_hk_ios/mappers/health_record_mappers/FloorsClimbedRecordMappers.swift
@@ -1,0 +1,67 @@
+import Foundation
+import HealthKit
+
+// ==================== FLOORS CLIMBED RECORD MAPPERS ====================
+
+extension FloorsClimbedRecordDto {
+    /**
+     * Converts this DTO to a HealthKit `HKQuantitySample`.
+     *
+     * - Throws: An error if the quantity type cannot be created.
+     */
+
+    func toHealthKit() throws -> HKQuantitySample {
+        guard let type = HKQuantityType.quantityType(forIdentifier: .flightsClimbed) else {
+            throw NSError(
+                domain: "HealthConnectorError",
+                code: -1,
+                userInfo: [NSLocalizedDescriptionKey: "Failed to create flights climbed quantity type"]
+            )
+        }
+
+        let quantity = HKQuantity(unit: .count(), doubleValue: floors.value)
+        let startDate = Date(timeIntervalSince1970: TimeInterval(startTime) / 1000.0)
+        let endDate = Date(timeIntervalSince1970: TimeInterval(endTime) / 1000.0)
+
+        return HKQuantitySample(
+            type: type,
+            quantity: quantity,
+            start: startDate,
+            end: endDate,
+            device: metadata.toHealthKitDevice(),
+            metadata: metadata.toHealthKitMetadata(timeZone: TimeZone.current)
+        )
+    }
+}
+
+extension HKQuantitySample {
+    /**
+     * Converts this HealthKit sample to a `FloorsClimbedRecordDto`.
+     *
+     * Returns `nil` if this sample is not a flights climbed sample.
+     */
+
+    func toFloorsClimbedRecordDto() -> FloorsClimbedRecordDto? {
+        guard quantityType.identifier == HKQuantityTypeIdentifier.flightsClimbed.rawValue else {
+            return nil
+        }
+
+        let unit = HKUnit.count()
+        let value = quantity.doubleValue(for: unit)
+
+        let metadataDict = metadata ?? [:]
+        let zoneOffset = metadataDict.extractTimeZoneOffset(for: startDate)
+        
+        return FloorsClimbedRecordDto(
+            floors: NumericDto(unit: NumericUnitDto.numeric, value: value),
+            endTime: Int64(endDate.timeIntervalSince1970 * 1000),
+            id: uuid.uuidString,
+            metadata: metadataDict.toMetadataDto(
+                source: sourceRevision.source,
+                device: device
+            ),
+            startTime: Int64(startDate.timeIntervalSince1970 * 1000),
+            zoneOffsetSeconds: zoneOffset
+        )
+    }
+}
