+++ packages/health_connector_hk_ios/ios/health_connector_hk_ios/Sources/health_connector_hk_ios/mappers/health_record_mappers/LeanBodyMassRecordMappers.swift
@@ -1,0 +1,62 @@
+import Foundation
+import HealthKit
+
+// ==================== LEAN BODY MASS RECORD MAPPERS ====================
+
+extension LeanBodyMassRecordDto {
+    /**
+     * Converts this DTO to a HealthKit `HKQuantitySample`.
+     *
+     * - Throws: An error if the quantity type cannot be created.
+     */
+
+    func toHealthKit() throws -> HKQuantitySample {
+        guard let type = HKQuantityType.quantityType(forIdentifier: .leanBodyMass) else {
+            throw NSError(
+                domain: "HealthConnectorError",
+                code: -1,
+                userInfo: [NSLocalizedDescriptionKey: "Failed to create lean body mass quantity type"]
+            )
+        }
+
+        let quantity = mass.toHealthKit()
+        let date = Date(timeIntervalSince1970: TimeInterval(time) / 1000.0)
+
+        return HKQuantitySample(
+            type: type,
+            quantity: quantity,
+            start: date,
+            end: date, // Instant records have same start and end
+            device: metadata.toHealthKitDevice(),
+            metadata: metadata.toHealthKitMetadata()
+        )
+    }
+}
+
+extension HKQuantitySample {
+    /**
+     * Converts this HealthKit sample to a `LeanBodyMassRecordDto`.
+     *
+     * Returns `nil` if this sample is not a lean body mass sample.
+     */
+
+    func toLeanBodyMassRecordDto() -> LeanBodyMassRecordDto? {
+        guard quantityType.identifier == HKQuantityTypeIdentifier.leanBodyMass.rawValue else {
+            return nil
+        }
+
+        let metadataDict = metadata ?? [:]
+        let zoneOffset = metadataDict.extractTimeZoneOffset(for: startDate)
+        
+        return LeanBodyMassRecordDto(
+            id: uuid.uuidString,
+            metadata: metadataDict.toMetadataDto(
+                source: sourceRevision.source,
+                device: device
+            ),
+            time: Int64(startDate.timeIntervalSince1970 * 1000),
+            mass: quantity.toMassDto(),
+            zoneOffsetSeconds: zoneOffset
+        )
+    }
+}
