+++ packages/health_connector_hk_ios/ios/health_connector_hk_ios/Sources/health_connector_hk_ios/mappers/health_record_mappers/StepsRecordMappers.swift
@@ -1,0 +1,66 @@
+import Foundation
+import HealthKit
+
+// ==================== STEP RECORD MAPPERS ====================
+
+extension StepRecordDto {
+    /**
+     * Converts this DTO to a HealthKit `HKQuantitySample`.
+     *
+     * - Throws: An error if the quantity type cannot be created.
+     */
+
+    func toHealthKit() throws -> HKQuantitySample {
+        guard let type = HKQuantityType.quantityType(forIdentifier: .stepCount) else {
+            throw NSError(
+                domain: "HealthConnectorError",
+                code: -1,
+                userInfo: [NSLocalizedDescriptionKey: "Failed to create step count quantity type"]
+            )
+        }
+
+        let quantity = HKQuantity(unit: .count(), doubleValue: count.value)
+        let startDate = Date(timeIntervalSince1970: TimeInterval(startTime) / 1000.0)
+        let endDate = Date(timeIntervalSince1970: TimeInterval(endTime) / 1000.0)
+
+        return HKQuantitySample(
+            type: type,
+            quantity: quantity,
+            start: startDate,
+            end: endDate,
+            device: metadata.toHealthKitDevice(),
+            metadata: metadata.toHealthKitMetadata(timeZone: TimeZone.current)
+        )
+    }
+}
+
+extension HKQuantitySample {
+    /**
+     * Converts this HealthKit sample to a `StepRecordDto`.
+     *
+     * Returns `nil` if this sample is not a step count sample.
+     */
+
+    func toStepRecordDto() -> StepRecordDto? {
+        guard quantityType.identifier == HKQuantityTypeIdentifier.stepCount.rawValue else {
+            return nil
+        }
+
+        let count = Int64(quantity.doubleValue(for: .count()))
+
+        let metadataDict = metadata ?? [:]
+        let zoneOffset = metadataDict.extractTimeZoneOffset(for: startDate)
+        
+        return StepRecordDto(
+            count: count.toNumericDto(),
+            endTime: Int64(endDate.timeIntervalSince1970 * 1000),
+            id: uuid.uuidString,
+            metadata: metadataDict.toMetadataDto(
+                source: sourceRevision.source,
+                device: device
+            ),
+            startTime: Int64(startDate.timeIntervalSince1970 * 1000),
+            zoneOffsetSeconds: zoneOffset
+        )
+    }
+}
