+++ packages/health_connector_hk_ios/lib/src/health_connector_hk_client.dart
@@ -21,7 +21,13 @@ import 'package:health_connector_core/health_connector_core.dart'
         PermissionListExtension,
         sinceV1_0_0,
         internalUse;
-import 'package:health_connector_hk_ios/src/mappers/mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/error_code_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/health_data_type_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/health_record_mapper.dart';
+import 'package:health_connector_hk_ios/src/mappers/health_record_mappers/health_record_id_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/permission_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/request_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/response_mappers.dart';
 import 'package:health_connector_hk_ios/src/pigeon/health_connector_platform_api.g.dart'
     show
         DeleteRecordsByIdsRequestDto,
