+++ packages/health_connector_hk_ios/lib/src/mappers/health_record_mappers/active_calories_burned_record_mappers.dart
@@ -1,0 +1,40 @@
+import 'package:health_connector_core/health_connector_core.dart'
+    show ActiveCaloriesBurnedRecord, Energy, HealthRecordId;
+import 'package:health_connector_hk_ios/src/mappers/health_record_mappers/health_record_id_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/measurement_unit_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/metadata_mappers.dart';
+import 'package:health_connector_hk_ios/src/pigeon/health_connector_platform_api.g.dart'
+    show ActiveCaloriesBurnedRecordDto, EnergyDto;
+import 'package:meta/meta.dart' show internal;
+
+/// Converts [ActiveCaloriesBurnedRecord] to [ActiveCaloriesBurnedRecordDto].
+@internal
+extension ActiveCaloriesBurnedRecordToDto on ActiveCaloriesBurnedRecord {
+  ActiveCaloriesBurnedRecordDto toDto() {
+    return ActiveCaloriesBurnedRecordDto(
+      id: id.toDto(),
+      startTime: startTime.millisecondsSinceEpoch,
+      endTime: endTime.millisecondsSinceEpoch,
+      zoneOffsetSeconds: startZoneOffsetSeconds ?? endZoneOffsetSeconds,
+      metadata: metadata.toDto(),
+      energy: energy.toDto() as EnergyDto,
+    );
+  }
+}
+
+/// Converts [ActiveCaloriesBurnedRecordDto] to [ActiveCaloriesBurnedRecord].
+@internal
+extension ActiveCaloriesBurnedRecordDtoToDomain
+    on ActiveCaloriesBurnedRecordDto {
+  ActiveCaloriesBurnedRecord toDomain() {
+    return ActiveCaloriesBurnedRecord(
+      id: id?.toDomain() ?? HealthRecordId.none,
+      startTime: DateTime.fromMillisecondsSinceEpoch(startTime),
+      endTime: DateTime.fromMillisecondsSinceEpoch(endTime),
+      startZoneOffsetSeconds: zoneOffsetSeconds,
+      endZoneOffsetSeconds: zoneOffsetSeconds,
+      metadata: metadata.toDomain(),
+      energy: energy.toDomain() as Energy,
+    );
+  }
+}
