+++ packages/health_connector_hk_ios/lib/src/mappers/health_record_mappers/body_fat_percentage_record_mappers.dart
@@ -1,0 +1,36 @@
+import 'package:health_connector_core/health_connector_core.dart'
+    show BodyFatPercentageRecord, Percentage, HealthRecordId;
+import 'package:health_connector_hk_ios/src/mappers/health_record_mappers/health_record_id_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/measurement_unit_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/metadata_mappers.dart';
+import 'package:health_connector_hk_ios/src/pigeon/health_connector_platform_api.g.dart'
+    show BodyFatPercentageRecordDto, PercentageDto;
+import 'package:meta/meta.dart' show internal;
+
+/// Converts [BodyFatPercentageRecord] to [BodyFatPercentageRecordDto].
+@internal
+extension BodyFatPercentageRecordToDto on BodyFatPercentageRecord {
+  BodyFatPercentageRecordDto toDto() {
+    return BodyFatPercentageRecordDto(
+      id: id.toDto(),
+      time: time.millisecondsSinceEpoch,
+      zoneOffsetSeconds: zoneOffsetSeconds,
+      metadata: metadata.toDto(),
+      percentage: percentage.toDto() as PercentageDto,
+    );
+  }
+}
+
+/// Converts [BodyFatPercentageRecordDto] to [BodyFatPercentageRecord].
+@internal
+extension BodyFatPercentageRecordDtoToDomain on BodyFatPercentageRecordDto {
+  BodyFatPercentageRecord toDomain() {
+    return BodyFatPercentageRecord(
+      id: id?.toDomain() ?? HealthRecordId.none,
+      time: DateTime.fromMillisecondsSinceEpoch(time),
+      zoneOffsetSeconds: zoneOffsetSeconds,
+      metadata: metadata.toDomain(),
+      percentage: percentage.toDomain() as Percentage,
+    );
+  }
+}
