+++ packages/health_connector_hk_ios/lib/src/mappers/health_record_mappers/body_temperature_record_mappers.dart
@@ -1,0 +1,36 @@
+import 'package:health_connector_core/health_connector_core.dart'
+    show BodyTemperatureRecord, Temperature, HealthRecordId;
+import 'package:health_connector_hk_ios/src/mappers/health_record_mappers/health_record_id_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/measurement_unit_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/metadata_mappers.dart';
+import 'package:health_connector_hk_ios/src/pigeon/health_connector_platform_api.g.dart'
+    show BodyTemperatureRecordDto, TemperatureDto;
+import 'package:meta/meta.dart' show internal;
+
+/// Converts [BodyTemperatureRecord] to [BodyTemperatureRecordDto].
+@internal
+extension BodyTemperatureRecordToDto on BodyTemperatureRecord {
+  BodyTemperatureRecordDto toDto() {
+    return BodyTemperatureRecordDto(
+      id: id.toDto(),
+      time: time.millisecondsSinceEpoch,
+      zoneOffsetSeconds: zoneOffsetSeconds,
+      metadata: metadata.toDto(),
+      temperature: temperature.toDto() as TemperatureDto,
+    );
+  }
+}
+
+/// Converts [BodyTemperatureRecordDto] to [BodyTemperatureRecord].
+@internal
+extension BodyTemperatureRecordDtoToDomain on BodyTemperatureRecordDto {
+  BodyTemperatureRecord toDomain() {
+    return BodyTemperatureRecord(
+      id: id?.toDomain() ?? HealthRecordId.none,
+      time: DateTime.fromMillisecondsSinceEpoch(time),
+      zoneOffsetSeconds: zoneOffsetSeconds,
+      metadata: metadata.toDomain(),
+      temperature: temperature.toDomain() as Temperature,
+    );
+  }
+}
