+++ packages/health_connector_hk_ios/lib/src/mappers/health_record_mappers/health_record_id_mappers.dart
@@ -1,0 +1,20 @@
+import 'package:health_connector_core/health_connector_core.dart'
+    show HealthRecordId;
+import 'package:meta/meta.dart' show internal;
+
+/// Converts [HealthRecordId] to [String] for DTO transfer.
+@internal
+extension HealthRecordIdFromDomainToDto on HealthRecordId {
+  String toDto() => value;
+}
+
+/// Converts [String] to [HealthRecordId].
+@internal
+extension HealthRecordIdFromDtoToDomain on String {
+  HealthRecordId toDomain() {
+    if (this == HealthRecordId.none.value) {
+      return HealthRecordId.none;
+    }
+    return HealthRecordId(this);
+  }
+}
