+++ packages/health_connector_hk_ios/lib/src/mappers/health_record_mappers/heart_rate_measurement_mappers.dart
@@ -1,0 +1,28 @@
+import 'package:health_connector_core/health_connector_core.dart'
+    show HeartRateMeasurement, Numeric;
+import 'package:health_connector_hk_ios/src/mappers/measurement_unit_mappers.dart';
+import 'package:health_connector_hk_ios/src/pigeon/health_connector_platform_api.g.dart'
+    show HeartRateMeasurementDto, NumericDto;
+import 'package:meta/meta.dart' show internal;
+
+/// Converts [HeartRateMeasurement] to [HeartRateMeasurementDto].
+@internal
+extension HeartRateMeasurementDomainToDto on HeartRateMeasurement {
+  HeartRateMeasurementDto toDto() {
+    return HeartRateMeasurementDto(
+      time: time.millisecondsSinceEpoch,
+      beatsPerMinute: beatsPerMinute.toDto() as NumericDto,
+    );
+  }
+}
+
+/// Converts [HeartRateMeasurementDto] to [HeartRateMeasurement].
+@internal
+extension HeartRateMeasurementDtoToDomain on HeartRateMeasurementDto {
+  HeartRateMeasurement toDomain() {
+    return HeartRateMeasurement(
+      time: DateTime.fromMillisecondsSinceEpoch(time),
+      beatsPerMinute: beatsPerMinute.toDomain() as Numeric,
+    );
+  }
+}
