+++ packages/health_connector_hk_ios/lib/src/mappers/health_record_mappers/step_record_mappers.dart
@@ -1,0 +1,40 @@
+import 'package:health_connector_core/health_connector_core.dart'
+    show StepRecord, Numeric, HealthRecordId;
+import 'package:health_connector_hk_ios/src/mappers/health_record_mappers/health_record_id_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/measurement_unit_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/metadata_mappers.dart';
+import 'package:health_connector_hk_ios/src/pigeon/health_connector_platform_api.g.dart'
+    show StepRecordDto, NumericDto;
+import 'package:meta/meta.dart' show internal;
+
+/// Converts [StepRecord] to [StepRecordDto].
+@internal
+extension StepRecordToDto on StepRecord {
+  StepRecordDto toDto() {
+    return StepRecordDto(
+      id: id.toDto(),
+      startTime: startTime.millisecondsSinceEpoch,
+      endTime: endTime.millisecondsSinceEpoch,
+      zoneOffsetSeconds: startZoneOffsetSeconds,
+
+      metadata: metadata.toDto(),
+      count: count.toDto() as NumericDto,
+    );
+  }
+}
+
+/// Converts [StepRecordDto] to [StepRecord].
+@internal
+extension StepRecordDtoToDomain on StepRecordDto {
+  StepRecord toDomain() {
+    return StepRecord(
+      id: id?.toDomain() ?? HealthRecordId.none,
+      startTime: DateTime.fromMillisecondsSinceEpoch(startTime),
+      endTime: DateTime.fromMillisecondsSinceEpoch(endTime),
+      startZoneOffsetSeconds: zoneOffsetSeconds,
+      endZoneOffsetSeconds: zoneOffsetSeconds,
+      metadata: metadata.toDomain(),
+      count: count.toDomain() as Numeric,
+    );
+  }
+}
