+++ packages/health_connector_hk_ios/lib/src/mappers/health_record_mappers/weight_record_mappers.dart
@@ -1,0 +1,36 @@
+import 'package:health_connector_core/health_connector_core.dart'
+    show WeightRecord, Mass, HealthRecordId;
+import 'package:health_connector_hk_ios/src/mappers/health_record_mappers/health_record_id_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/measurement_unit_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/metadata_mappers.dart';
+import 'package:health_connector_hk_ios/src/pigeon/health_connector_platform_api.g.dart'
+    show WeightRecordDto, MassDto;
+import 'package:meta/meta.dart' show internal;
+
+/// Converts [WeightRecord] to [WeightRecordDto].
+@internal
+extension WeightRecordToDto on WeightRecord {
+  WeightRecordDto toDto() {
+    return WeightRecordDto(
+      id: id.toDto(),
+      time: time.millisecondsSinceEpoch,
+      zoneOffsetSeconds: zoneOffsetSeconds,
+      metadata: metadata.toDto(),
+      weight: weight.toDto() as MassDto,
+    );
+  }
+}
+
+/// Converts [WeightRecordDto] to [WeightRecord].
+@internal
+extension WeightRecordDtoToDomain on WeightRecordDto {
+  WeightRecord toDomain() {
+    return WeightRecord(
+      id: id?.toDomain() ?? HealthRecordId.none,
+      time: DateTime.fromMillisecondsSinceEpoch(time),
+      zoneOffsetSeconds: zoneOffsetSeconds,
+      metadata: metadata.toDomain(),
+      weight: weight.toDomain() as Mass,
+    );
+  }
+}
