+++ packages/health_connector_hk_ios/lib/src/mappers/health_record_mappers/wheelchair_pushes_record_mappers.dart
@@ -1,0 +1,40 @@
+import 'package:health_connector_core/health_connector_core.dart'
+    show WheelchairPushesRecord, Numeric, HealthRecordId;
+import 'package:health_connector_hk_ios/src/mappers/health_record_mappers/health_record_id_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/measurement_unit_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/metadata_mappers.dart';
+import 'package:health_connector_hk_ios/src/pigeon/health_connector_platform_api.g.dart'
+    show WheelchairPushesRecordDto, NumericDto;
+import 'package:meta/meta.dart' show internal;
+
+/// Converts [WheelchairPushesRecord] to [WheelchairPushesRecordDto].
+@internal
+extension WheelchairPushesRecordToDto on WheelchairPushesRecord {
+  WheelchairPushesRecordDto toDto() {
+    return WheelchairPushesRecordDto(
+      id: id.toDto(),
+      startTime: startTime.millisecondsSinceEpoch,
+      endTime: endTime.millisecondsSinceEpoch,
+      zoneOffsetSeconds: startZoneOffsetSeconds,
+
+      metadata: metadata.toDto(),
+      pushes: pushes.toDto() as NumericDto,
+    );
+  }
+}
+
+/// Converts [WheelchairPushesRecordDto] to [WheelchairPushesRecord].
+@internal
+extension WheelchairPushesRecordDtoToDomain on WheelchairPushesRecordDto {
+  WheelchairPushesRecord toDomain() {
+    return WheelchairPushesRecord(
+      id: id?.toDomain() ?? HealthRecordId.none,
+      startTime: DateTime.fromMillisecondsSinceEpoch(startTime),
+      endTime: DateTime.fromMillisecondsSinceEpoch(endTime),
+      startZoneOffsetSeconds: zoneOffsetSeconds,
+      endZoneOffsetSeconds: zoneOffsetSeconds,
+      metadata: metadata.toDomain(),
+      pushes: pushes.toDomain() as Numeric,
+    );
+  }
+}
