--- packages/health_connector_hk_ios/lib/src/mappers/mappers.dart
@@ -1,12 +1,0 @@
-library;
-
-export 'aggregation_metric_mappers.dart';
-export 'error_code_mappers.dart';
-export 'health_data_mappers.dart';
-export 'health_record_mappers.dart';
-export 'meal_type_mappers.dart';
-export 'measurement_unit_mappers.dart';
-export 'metadata_mappers.dart';
-export 'permission_mappers.dart';
-export 'request_mappers.dart';
-export 'response_mappers.dart';
