+++ packages/health_connector_hk_ios/lib/src/mappers/permission_mappers.dart
@@ -9,7 +9,7 @@ import 'package:health_connector_core/health_connector_core.dart'
         NutritionHealthDataType,
         PermissionStatus,
         PermissionRequestResult;
-import 'package:health_connector_hk_ios/src/mappers/health_data_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/health_data_type_mappers.dart';
 import 'package:health_connector_hk_ios/src/pigeon/health_connector_platform_api.g.dart'
     show
         PermissionsRequestDto,
