+++ packages/health_connector_hk_ios/lib/src/mappers/request_mappers.dart
@@ -7,10 +7,9 @@ import 'package:health_connector_core/health_connector_core.dart'
         ReadRecordsRequest;
 import 'package:health_connector_hk_ios/src/mappers/'
     'aggregation_metric_mappers.dart';
-import 'package:health_connector_hk_ios/src/mappers/'
-    'health_data_mappers.dart';
-import 'package:health_connector_hk_ios/src/mappers/'
-    'health_record_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/health_data_type_mappers.dart';
+import 'package:health_connector_hk_ios/src/mappers/health_record_mapper.dart';
+import 'package:health_connector_hk_ios/src/mappers/health_record_mappers/health_record_id_mappers.dart';
 import 'package:health_connector_hk_ios/src/pigeon/health_connector_platform_api.g.dart'
     show
         ReadRecordRequestDto,
